pyproject.toml
[project]
name = "animal-classification-by-cnn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.7",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "scikit-learn>=1.7.2",
    "torch>=2.9.0",
    "torchsummary>=1.5.1",
    "torchvision>=0.24.0",
    "tqdm>=4.67.1",
]
